FoundGem/requirements.txt
flask==2.3.2
flask-login==0.6.2
flask-sqlalchemy==3.0.3
flask-wtf==1.2.1
python-dotenv==1.0.0
PyMySQL==1.0.2
cryptography==39.0.1
werkzeug==2.3.7
email-validator==2.0.0